+++ ethpm/_utils/chains.py
@@ -45,25 +45,6 @@ def is_BIP122_block_uri(value: URI) -> bool:
     return resource_type == BLOCK
 
 
-@curry
-def check_if_chain_matches_chain_uri(web3: Web3, blockchain_uri: URI) -> bool:
-    chain_id, resource_type, resource_hash = parse_BIP122_uri(blockchain_uri)
-    genesis_block = web3.eth.getBlock("earliest")
-
-    if encode_hex(genesis_block["hash"]) != chain_id:
-        return False
-
-    if resource_type == BLOCK:
-        resource = web3.eth.getBlock(resource_hash)
-    else:
-        raise ValueError(f"Unsupported resource type: {resource_type}")
-
-    if encode_hex(resource["hash"]) == resource_hash:
-        return True
-    else:
-        return False
-
-
 BLOCK_OR_TRANSACTION_HASH_REGEX = "^(?:0x)?[a-zA-Z0-9]{64}$"
 
 
@@ -101,21 +82,3 @@ def create_BIP122_uri(
 
 def create_block_uri(chain_id: str, block_identifier: str) -> URI:
     return create_BIP122_uri(chain_id, "block", remove_0x_prefix(block_identifier))
-
-
-def create_latest_block_uri(w3: Web3, from_blocks_ago: int = 3) -> URI:
-    """
-    Creates a block uri for the given w3 instance.
-    Defaults to 3 blocks prior to the "latest" block to accommodate for block reorgs.
-    If using a testnet with less than 3 mined blocks, adjust :from_blocks_ago:.
-    """
-    chain_id = to_hex(get_genesis_block_hash(w3))
-    latest_block_tx_receipt = w3.eth.getBlock("latest")
-    target_block_number = latest_block_tx_receipt.number - from_blocks_ago
-    if target_block_number < 0:
-        raise Exception(
-            f"Only {latest_block_tx_receipt.number} blocks avaible on provided w3, "
-            f"cannot create latest block uri for {from_blocks_ago} blocks ago."
-        )
-    recent_block = to_hex(w3.eth.getBlock(target_block_number).hash)
-    return create_block_uri(chain_id, recent_block)
