+++ ethpm/_utils/ipfs.py
@@ -7,7 +7,7 @@ from base58 import b58encode
 from eth_utils import to_text
 from google.protobuf.descriptor import Descriptor
 
-from ethpm.utils.protobuf.ipfs_file_pb2 import Data, PBNode
+from ethpm._utils.protobuf.ipfs_file_pb2 import Data, PBNode
 
 
 def dummy_ipfs_pin(path: Path) -> Dict[str, str]:
