+++ ethpm/tools/builder.py
@@ -22,7 +22,7 @@ from ethpm._utils.mappings import deep_merge_dicts
 from ethpm.backends.ipfs import BaseIPFSBackend
 from ethpm.exceptions import ManifestBuildingError, ValidationError
 from ethpm.package import format_manifest
-from ethpm.utils.chains import is_BIP122_block_uri
+from ethpm._utils.chains import is_BIP122_block_uri
 from ethpm.validation.manifest import (
     validate_manifest_against_schema,
 )
